Improve efficiency and clean code
Removed tedious line-by-line loops that parse the entry input files. Instead I treat most of the input as non-tex except
"/" "{" and "}". This allows for the command \entrytitle{My title here} to exist in a journal entry and be parsed correctly,
while still allowing other special characters to exist within the journal entry without breaking at compile-time. This is
much faster as TeX doesn't have to perform pattern matching for each line in the journal input files.